test(agent): split server exec tests by concern
Preserve all 37 existing tests while separating message, websocket I/O, resize queue, and handler coverage. Keep the websocket fixture in a shared test helper.

Fixes #23218
